Rename command_object to command_interface
